A working version of gpu forall offload.
This changeset incorporates an almost working version of gpu forall offload
capablities. One the opencl headers and typedefs are manually fixed, simple
forall loops can be successfully executed on the GPU.

Changes:
1. Remove GPU-based flags made redundant by moving the GPU offload pass to
much later in the compilation phase.
2. Rename FLAG_GPU_ON to FLAG_OFFLOAD_TO_GPU.
3. Estimate Trip count and replace loop indices with an apropriately scaled
and displaced version of the work-item id
4. Insert parentheses in GPUForLoop codegen
5. Rename GPUForLoop methods to denote the possiblity that the compiler does
not alwasy succeed in converting an unordered CForLoop to a GPUForLoop.
6. Move the createGPU almost to the very end - just before the "AST to
C/LLVM" passes.
7. Replace native vectors with STL vectors.
8. Fix external symbol capture and flattening for GPU offloadable functions.
9. Pass work-item count (aka trip-count) instead of work-group count
to the hsa_enqueue function.